refactor: Move enum validation to domain layer and update MemberEntity
